TechDraw: extract QGIDatumLabel
